rewrite to SDL2, because some linux distros doesn't have SDL3 in repositories yet

+++ src/gui.h
@@ -48,7 +48,7 @@ class Gui {
 		/// \endcond
 	private:
 		SDL_Window* window;
-		float mouse_x, mouse_y;
+		int mouse_x, mouse_y;
 		bool left_mouse_button_down;
 };
